feat: add the shortenAddress
