+++ src/utils/index.js
@@ -1,22 +1,6 @@
-import {isHexAddress} from '@cfxjs/account'
-import {validateBase32Address} from '@cfxjs/base32-address'
-import {ChainIdCfx} from '../constants/chainConfig'
-
 export const IS_DEV =
   window.location.hostname === 'localhost' ||
   window.location.hostname.indexOf('test') > -1
-
-export function checkHexAddress(address) {
-  return isHexAddress(address)
-}
-
-export function checkCfxTokenAddress(address) {
-  return validateBase32Address(
-    address,
-    IS_DEV ? ChainIdCfx.TESTNET : ChainIdCfx.MAINNET,
-    'contract',
-  )
-}
 
 export const getEllipsStr = (str, frontNum, endNum) => {
   if (str) {
@@ -31,16 +15,4 @@ export const getEllipsStr = (str, frontNum, endNum) => {
     )
   }
   return ''
-}
-
-export function shortenEthAddress(address) {
-  if (!checkHexAddress(address)) return ''
-  return getEllipsStr(address, 6, 4)
-}
-
-export function shortenCfxAddress(address) {
-  if (!checkCfxTokenAddress(address)) ''
-  const arr = address.split(':')
-  const secondStr = getEllipsStr(arr[1], 4, 4)
-  return arr[0] + ':' + secondStr
 }
